fix(layout): fix common layout
