+++ src/app/[locale]/(auth)/explore-huber/page.tsx
@@ -1,5 +1,11 @@
 const ExploreHuber = () => {
-  return <div>ExploreHuber</div>;
+  return (
+    <div className="px-28 py-12">
+      <div className="mx-auto h-full w-full max-w-[1216px] rounded-lg">
+        Hubers
+      </div>
+    </div>
+  );
 };
 
 export default ExploreHuber;
